Improve offline operations

+++ app/src/main/java/com/uva/inertia/besilite/SettingsFragment.java
@@ -1,11 +1,14 @@
 package com.uva.inertia.besilite;
 
+import android.app.AlertDialog;
 import android.content.Context;
+import android.content.DialogInterface;
 import android.os.Bundle;
 import android.os.Environment;
 import android.preference.Preference;
 import android.preference.PreferenceFragment;
 import android.util.Log;
+import android.widget.Toast;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -35,34 +38,54 @@ public class SettingsFragment extends PreferenceFragment {
         b.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
             @Override
             public boolean onPreferenceClick(Preference preference) {
-                File folder = new File(getActivity().getFilesDir(), "survey");
-                if (!folder.mkdirs()) {
-                    Log.e("FILES", "Did not create folder");
-                }
-                FileInputStream fin;
-                FileOutputStream fou;
 
-                Iterator<File> filesIt = Arrays.asList(folder.listFiles()).iterator();
-                for (; filesIt.hasNext();){
-                    if (isExternalStorageWritable()){
-                        try {
-                            //As per http://stackoverflow.com/questions/9292954/how-to-make-a-copy-of-a-file-in-android
-                            File f = filesIt.next();
-                            fin = new FileInputStream(f);
-                            File out = new File(Environment.getExternalStoragePublicDirectory("BESI"),f.getName()+".json");
-                            fou = new FileOutputStream(out);
-                            FileChannel inChan = fin.getChannel();
-                            FileChannel outChan = fou.getChannel();
-                            inChan.transferTo(0, inChan.size(), outChan);
-                            inChan.close();
-                            outChan.close();
-                            f.delete();
-                            Log.d("FILES","Exported: " + out.getName());
-                        } catch (IOException e) {
-                            e.printStackTrace();
+                DialogInterface.OnClickListener action = new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which) {
+                        if (which == DialogInterface.BUTTON_POSITIVE) {
+                            Toast.makeText(getActivity(),"Exporting files!", Toast.LENGTH_LONG).show();
+                            File folder = new File(getActivity().getFilesDir(), "offline");
+                            if (!folder.mkdirs()) {
+                                Log.e("FILES", "Did not create folder");
+                            }
+                            FileInputStream fin;
+                            FileOutputStream fou;
+
+                            Iterator<File> filesIt = Arrays.asList(folder.listFiles()).iterator();
+                            for (; filesIt.hasNext(); ) {
+                                if (isExternalStorageWritable()) {
+                                    try {
+                                        //As per http://stackoverflow.com/questions/9292954/how-to-make-a-copy-of-a-file-in-android
+                                        File f = filesIt.next();
+                                        fin = new FileInputStream(f);
+                                        File out = new File(Environment.getExternalStoragePublicDirectory("BESI"), f.getName() + ".json");
+                                        fou = new FileOutputStream(out);
+                                        FileChannel inChan = fin.getChannel();
+                                        FileChannel outChan = fou.getChannel();
+                                        inChan.transferTo(0, inChan.size(), outChan);
+                                        inChan.close();
+                                        outChan.close();
+                                        f.delete();
+                                        Log.d("FILES", "Exported: " + out.getName());
+                                    } catch (IOException e) {
+                                        e.printStackTrace();
+                                    }
+                                }
+                            }
                         }
                     }
-                }
+                };
+
+                AlertDialog.Builder diaBuild = new AlertDialog.Builder(getActivity())
+                        .setMessage("This action will delete all offline items from internal storage. Are you sure you want to continue?")
+                        .setNegativeButton("No", action)
+                        .setPositiveButton("Yes", action);
+
+                diaBuild.show();
+
+
+
+
 
 
                 return true;
